Fix Notification message
