+++ src/main/java/de/unistuttgart/iste/meitrex/course_service/service/ChapterService.java
@@ -198,7 +198,7 @@ public class ChapterService {
         for (ChapterEntity ch : unlockList) {
             String CourseTitle = courseService.getCourseById(ch.getCourseId()).getTitle();
             List<UUID> userIds = membershipService.getUserIdsOfCourse(ch.getCourseId());
-            topicPublisher.notificationEvent(ch.getCourseId(), userIds, ServerSource.CHAPTER, "/courses/" + ch.getCourseId() + "/chapters/" + ch.getId(), "New Chapter Unlocked!", CourseTitle + "'s Chapter" + ch.getTitle() + "is unlocked");
+            topicPublisher.notificationEvent(ch.getCourseId(), userIds, ServerSource.CHAPTER, "/courses/" + ch.getCourseId() + "/chapters/" + ch.getId(), "New Chapter Unlocked!", CourseTitle + "'s Chapter" + ch.getTitle() + " is unlocked");
         }
 
         // Lock
@@ -206,7 +206,7 @@ public class ChapterService {
         for (ChapterEntity ch : lockList) {
             String CourseTitle = courseService.getCourseById(ch.getCourseId()).getTitle();
             List<UUID> userIds = membershipService.getUserIdsOfCourse(ch.getCourseId());
-            topicPublisher.notificationEvent(ch.getCourseId(), userIds, ServerSource.CHAPTER, "/courses/" + ch.getCourseId() + "/chapters/" + ch.getId(), "An Old Chapter Locked", CourseTitle + "'s Chapter" + ch.getTitle() + "is locked");
+            topicPublisher.notificationEvent(ch.getCourseId(), userIds, ServerSource.CHAPTER, "/courses/" + ch.getCourseId() + "/chapters/" + ch.getId(), "An Old Chapter Locked", CourseTitle + "'s Chapter" + ch.getTitle() + " is locked");
         }
     }
 }
