+++ src/test/java/de/unistuttgart/iste/meitrex/course_service/service/CourseServiceTest.java
@@ -321,43 +321,4 @@ class CourseServiceTest {
                 .build();
     }
 
-    @Test void unlockSendsNotification() {
-        var repo=mock(ChapterRepository.class);
-        var map=mock(ChapterMapper.class);
-        var course=mock(CourseService.class);
-        var ms=mock(MembershipService.class);
-        var val=mock(ChapterValidator.class);
-        var pub=mock(TopicPublisher.class);
-        var s=new ChapterService(map,repo,course,ms,val,pub);
-        UUID cid=UUID.randomUUID(), chid=UUID.randomUUID();
-        var ch=ChapterEntity.builder().id(chid).courseId(cid).title("Intro").build();
-        when(repo.findChaptersToUnlock(any())).thenReturn(List.of(ch));
-        when(repo.findChaptersToLock(any())).thenReturn(List.of());
-        var dto=new Course(); dto.setTitle("CS101"); when(course.getCourseById(cid)).thenReturn(dto);
-        var users=List.of(UUID.randomUUID()); when(ms.getUserIdsOfCourse(cid)).thenReturn(users);
-        s.checkChapters();
-        verify(pub).notificationEvent(eq(cid), eq(users), eq(ServerSource.CHAPTER),
-                eq("/courses/"+cid+"/chapters/"+chid), eq("New Chapter Unlocked!"),
-                eq("CS101's Chapter Intro is unlocked"));
-    }
-
-    @Test void lockSendsNotification() {
-        var repo=mock(ChapterRepository.class);
-        var map=mock(ChapterMapper.class);
-        var course=mock(CourseService.class);
-        var ms=mock(MembershipService.class);
-        var val=mock(ChapterValidator.class);
-        var pub=mock(TopicPublisher.class);
-        var s=new ChapterService(map,repo,course,ms,val,pub);
-        UUID cid=UUID.randomUUID(), chid=UUID.randomUUID();
-        var ch=ChapterEntity.builder().id(chid).courseId(cid).title("Intro").build();
-        when(repo.findChaptersToUnlock(any())).thenReturn(List.of());
-        when(repo.findChaptersToLock(any())).thenReturn(List.of(ch));
-        var dto=new Course(); dto.setTitle("CS101"); when(course.getCourseById(cid)).thenReturn(dto);
-        var users=List.of(UUID.randomUUID()); when(ms.getUserIdsOfCourse(cid)).thenReturn(users);
-        s.checkChapters();
-        verify(pub).notificationEvent(eq(cid), eq(users), eq(ServerSource.CHAPTER),
-                eq("/courses/"+cid+"/chapters/"+chid), eq("An Old Chapter Locked"),
-                eq("CS101's Chapter Intro is locked"));
-    }
 }
